specdep: Rename to planex-depend
